+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - Book: Recommend using the `-Zunstable-options --fail-fast` argument to test targets to speed up mutation testing, on recent nightly toolchains.
 
+- Fixed: Don't error if the `--in-diff` patch file contains non-UTF-8 data in non-Rust files.
+
 ## 25.3.1 2025-08-10
 
 - Fixed: cargo-mutants' own tests were failing on nightly due to a change in the format of messages emitted by tests.